Create README.md
The data is for 2016-17 NBA season. The NBA season starts in October and ended in June. The GSW (Golden State Warriors) won the NBA finals on June 12, 2017 by 4-1 in the best of 7-game series beating Cleveland Cavaliers. Note that there are two rows for each game. For example, the row 2 and row 24 are both for the same game which was played on October 26, 2016 between Miami and Orlando. The statistics provided in columns K through AB are for the team listed in Column A. Meaning, you will need both rows for each game to get team statistics for both teams. Also, the TeamScr (Column G) is the points scored by team listed in Column A. And, the OppScr (Column H) has the points scored by team listed in Column D.